move api version constants and add explanatory comment
